separate hook from fucntions
